Add bluprints
